+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added missing x-label to the pvalue histogram.
 - Fixed wrong y-label in volcano-ma plot.
 - Fixed issue of wrong column renaming when calling `Qtable.make_sample_table` with `samples_as_columns=True` when the design contained sample names that are substrings of other sample names.
+- Fixed faulty modification site probability extraction for peptides with multiple modifications from FragPipe results, which was caused by a change in the site localization format in FragPipe version 22.0
 
 ----------------------------------------------------------------------------------------